refactor(test): playwrightテストの共通部分を少しリファクタリング、歌詞編集テストの移動
